[UPDATE] Change DB Schema
